refactor: Use BigDecimal for Order totalPrice

+++ domain/src/main/java/hu/vibe/homework/order/domain/Order.java
@@ -1,5 +1,6 @@
 package hu.vibe.homework.order.domain;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
@@ -8,7 +9,7 @@ public record Order(
         UUID id,
         UUID customerId,
         Instant createdAt,
-        double totalPrice,
+        BigDecimal totalPrice,
         List<OrderItem> items,
         OrderStatus status,
         Address shippingAddress,
