+++ domain/src/main/java/hu/vibe/homework/order/domain/OrderItem.java
@@ -1,3 +1,5 @@
 package hu.vibe.homework.order.domain;
 
-public record OrderItem(String productCode, int quantity, double unitPrice) {}
+import java.math.BigDecimal;
+
+public record OrderItem(String productCode, int quantity, BigDecimal unitPrice) {}
